# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Guidr Education Hub API Specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/signup:
    post:
      operationId: signupStudent
      tags: [auth]
      summary: Register a new student user with 3-step profile details
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupRequest"
      responses:
        "201":
          description: Student successfully registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUserResponse"
        "400":
          description: Validation error or email already exists
  /auth/login:
    post:
      operationId: loginUser
      tags: [auth]
      summary: Login for student or admin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Successfully authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUserResponse"
        "401":
          description: Invalid credentials
...
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get currently authenticated user and profile
      responses:
        "200":
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUserResponse"
        "401":
          description: Not authenticated
...
  /lessons/{id}:
    get:
      operationId: getLessonById
      tags: [lessons]
      summary: Get single lesson details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Lesson details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LessonDetail"
  /lessons/{id}/progress:
    post:
      operationId: toggleLessonProgress
      tags: [lessons]
      summary: Mark lesson completed or in progress
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                completed:
                  type: boolean
              required: [completed]
      responses:
        "200":
          description: Progress updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /lessons/{id}/chat:
    post:
      operationId: sendLessonChatMessage
      tags: [tutor]
      summary: Send question to Guidr Tutor AI RAG assistant for a specific lesson
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                message:
                  type: string
              required: [message]
      responses:
        "200":
          description: AI Tutor response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TutorChatResponse"
  /lessons/{id}/chat/history:
    get:
      operationId: getLessonChatHistory
      tags: [tutor]
      summary: Get chat conversation history for a specific lesson
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of chat messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatMessage"
...
  /admin/lessons:
    get:
      operationId: adminListLessons
      tags: [admin-lessons]
      summary: List all lessons for admin
      responses:
        "200":
          description: List of all lessons
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lesson"
    post:
      operationId: adminCreateLesson
      tags: [admin-lessons]
      summary: Create new lesson and trigger RAG chunking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLessonRequest"
      responses:
        "201":
          description: Created lesson
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lesson"
  /admin/lessons/{id}:
    put:
      operationId: adminUpdateLesson
      tags: [admin-lessons]
      summary: Update lesson content, status, module
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLessonRequest"
      responses:
        "200":
          description: Updated lesson
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lesson"
    delete:
      operationId: adminDeleteLesson
      tags: [admin-lessons]
      summary: Delete lesson
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted lesson
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
    AuthUserResponse:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [student, admin]
        studentProfile:
          $ref: "#/components/schemas/StudentProfile"
        adminProfile:
          $ref: "#/components/schemas/AdminProfile"
      required: [id, email, role]
    SignupRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
        school:
          type: string
        governorate:
          type: string
        grade:
          type: integer
        interests:
          type: array
          items:
            type: string
      required: [email, password, firstName, lastName, governorate, grade]
    LoginRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]
    StudentProfile:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
        school:
          type: string
        governorate:
          type: string
        grade:
          type: integer
        interests:
          type: array
          items:
            type: string
      required: [id, userId, firstName, lastName]
...
    AdminProfile:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        name:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [admin, super_admin]
      required: [id, userId, name, role]
...
    Lesson:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        module:
          type: string
        content:
          type: string
        status:
          type: string
          enum: [draft, published, locked]
        sortOrder:
          type: integer
        completed:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, title, module, status, sortOrder]
    LessonDetail:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        module:
          type: string
        content:
          type: string
        status:
          type: string
          enum: [draft, published, locked]
        sortOrder:
          type: integer
        completed:
          type: boolean
        chunksCount:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, title, module, content, status, sortOrder]
    CreateLessonRequest:
      type: object
      properties:
        title:
          type: string
        module:
          type: string
        content:
          type: string
        status:
          type: string
          enum: [draft, published, locked]
        sortOrder:
          type: integer
      required: [title, module, content, status]
    ChatMessage:
      type: object
      properties:
        id:
          type: string
        studentId:
          type: string
        lessonId:
          type: string
        role:
          type: string
          enum: [student, assistant]
        content:
          type: string
        createdAt:
          type: string
      required: [id, studentId, lessonId, role, content, createdAt]
    TutorChatResponse:
      type: object
      properties:
        response:
          type: string
        sources:
          type: array
          items:
            type: string
      required: [response]